requirements.txt
Django==1.7.4
django-custom-user==0.5
django-json-field==0.5.7
django-storages==1.1.8
djangorestframework==3.0.5
drf-nested-routers==0.9.0
python-dateutil==2.4.0
six==1.9.0
